requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bcrypt==3.2.2
cffi==2.0.0
click==8.2.1
dnspython==2.8.0
ecdsa==0.19.1
fastapi==0.116.1
h11==0.16.0
idna==3.10
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.14.1
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
